added float to python radius

+++ calculateCircleArea.py
@@ -2,4 +2,4 @@ import math
 
 print("Enter radius in metres: ")
 radius = input()
-print(math.pi * radius * radius)
+print(math.pi * float(radius * radius))
